third_party/WebKit/LayoutTests/platform/win7/external/wpt/html/dom/documents/resource-metadata-management/document-lastModified-01-expected.txt

This is a testharness.js-based test.
FAIL Date returned by lastModified is current at page load assert_approx_equals: expected 1528643301 +/- 2.5 but got 1528646900
PASS Date returned by lastModified is in the form "MM/DD/YYYY hh:mm:ss".
FAIL Date returned by lastModified is in the user's local time zone. assert_approx_equals: Hours and minutes should match local time. expected 25700 +/- 2 but got 29300
FAIL Date returned by lastModified is current after timeout. assert_approx_equals: (initial value was 1528646900) expected 1528643305 +/- 2.5 but got 1528646904
Harness: the test ran to completion.
